+++ src/playground/ec/.storybook/preview.js
@@ -17,6 +17,12 @@ addParameters({
       picked: false,
       hideCode: true,
     },
+    {
+      id: 'reset',
+      code: `<link rel="stylesheet" type="text/css" href="./styles/reset.css" />`,
+      picked: true,
+      hideCode: true,
+    },
   ],
   controls: { expanded: true },
   layout: 'fullscreen',
