--- src/presets/ec-preset-website/README.md
@@ -1,5 +1,0 @@
-# ECL preset: EC website
-
-This preset ships all the EC components with normalize.css and some styles applied to the `body`.
-
-This is the main preset if you're building an EC website with the ECL from the start.
