--- src/presets/ec-preset-website/src/ec-preset-website.js
@@ -1,3 +1,0 @@
-// Export components
-
-export * from '@ecl/preset-dev/src/dev';
