--- src/presets/ec-preset-website/src/ec-preset-website.scss
@@ -1,4 +1,0 @@
-@use 'normalize.css/normalize';
-@use 'globals';
-@use '@ecl/theme-ec-core';
-@use '@ecl/preset-dev/src/dev';
