+++ src/presets/reset/README.md
@@ -1,0 +1,5 @@
+# ECL preset: reset
+
+This preset contains a css reset (normalize.css) and some styles applied to the `body`.
+
+It only provides an optional css file.
