+++ src/presets/reset/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@ecl/preset-reset",
+  "author": "European Commission",
+  "license": "EUPL-1.1",
+  "version": "3.0.0",
+  "description": "ECL Reset preset",
+  "style": "dist/styles/reset.css",
+  "private": true,
+  "scripts": {
+    "build:copy": "ecl-builder copy",
+    "build:scripts": "ecl-builder scripts",
+    "build:styles": "ecl-builder styles",
+    "build": "run-p build:*",
+    "clean": "rimraf build dist",
+    "dist": "cross-env NODE_ENV=production run-p build:*",
+    "watch": "ecl-builder watch",
+    "prepublish": "npm run clean && run-p build dist"
+  },
+  "devDependencies": {
+    "@ecl/builder": "^3.0.0",
+    "chokidar-cli": "2.1.0",
+    "cross-env": "7.0.2",
+    "normalize.css": "8.0.1",
+    "npm-run-all": "4.1.5",
+    "rimraf": "3.0.2"
+  },
+  "keywords": [
+    "ecl",
+    "europa-component-library",
+    "design-system"
+  ],
+  "homepage": "https://github.com/ec-europa/europa-component-library",
+  "bugs": {
+    "url": "https://github.com/ec-europa/europa-component-library/issues"
+  }
+}
