+++ src/presets/reset/src/reset.scss
@@ -1,3 +1,5 @@
+@use 'normalize.css/normalize';
+
 body {
   margin: 0;
   overflow-x: hidden;
