+++ src/website/scripts/copy.sh
@@ -10,7 +10,5 @@ cp -r ../../dist/playground/ec ./public/playground/ec
 cp -r ../../dist/playground/eu ./public/playground/eu
 
 # Make assets available for storybook
-cp -r ../../dist/packages/ec-preset-website/. ./public/playground/ec
-
-cp -r ../../dist/packages/ec-core/. ./public/playground/ec
-cp -r ../../dist/packages/eu-core/. ./public/playground/eu
+cp -r ../../dist/packages/ec/. ./public/playground/ec
+cp -r ../../dist/packages/eu/. ./public/playground/eu
